# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  personnelId: zod.number().nullish(),
  name: zod.string().nullish(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user session
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  personnelId: zod.number().nullish(),
  name: zod.string().nullish(),
});

/**
 * @summary List all personnel
 */
export const ListPersonnelResponseItem = zod.object({
  id: zod.number(),
  lastName: zod.string(),
  firstName: zod.string(),
  middleInitial: zod.string().nullish(),
  employeeId: zod.string(),
  department: zod.string(),
  position: zod.string(),
  photoUrl: zod.string().nullish(),
  vehiclePlate: zod.string().nullish(),
  hasAccount: zod.boolean(),
  createdAt: zod.string(),
});
export const ListPersonnelResponse = zod.array(ListPersonnelResponseItem);

/**
 * @summary Create personnel and optionally an account
 */
export const CreatePersonnelBody = zod.object({
  lastName: zod.string(),
  firstName: zod.string(),
  middleInitial: zod.string().nullish(),
  employeeId: zod.string(),
  department: zod.string(),
  position: zod.string(),
  photoUrl: zod.string().nullish(),
  vehiclePlate: zod.string().nullish(),
  createAccount: zod.boolean().optional(),
  password: zod.string().nullish(),
});

/**
 * @summary Get a personnel by ID
 */
export const GetPersonnelParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPersonnelResponse = zod.object({
  id: zod.number(),
  lastName: zod.string(),
  firstName: zod.string(),
  middleInitial: zod.string().nullish(),
  employeeId: zod.string(),
  department: zod.string(),
  position: zod.string(),
  photoUrl: zod.string().nullish(),
  vehiclePlate: zod.string().nullish(),
  hasAccount: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Update personnel
 */
export const UpdatePersonnelParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePersonnelBody = zod.object({
  lastName: zod.string().optional(),
  firstName: zod.string().optional(),
  middleInitial: zod.string().nullish(),
  department: zod.string().optional(),
  position: zod.string().optional(),
  photoUrl: zod.string().nullish(),
  vehiclePlate: zod.string().nullish(),
});

export const UpdatePersonnelResponse = zod.object({
  id: zod.number(),
  lastName: zod.string(),
  firstName: zod.string(),
  middleInitial: zod.string().nullish(),
  employeeId: zod.string(),
  department: zod.string(),
  position: zod.string(),
  photoUrl: zod.string().nullish(),
  vehiclePlate: zod.string().nullish(),
  hasAccount: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete personnel
 */
export const DeletePersonnelParams = zod.object({
  id: zod.coerce.number(),
});

export const DeletePersonnelResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all accounts
 */
export const ListAccountsResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  personnelId: zod.number().nullish(),
  personnelName: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListAccountsResponse = zod.array(ListAccountsResponseItem);

/**
 * @summary Create account for existing personnel
 */
export const CreateAccountBody = zod.object({
  personnelId: zod.number(),
  password: zod.string(),
});

/**
 * @summary Update account password
 */
export const UpdateAccountParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAccountBody = zod.object({
  password: zod.string(),
});

export const UpdateAccountResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  personnelId: zod.number().nullish(),
  personnelName: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete account
 */
export const DeleteAccountParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAccountResponse = zod.object({
  message: zod.string(),
});