# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountRole } from "./accountRole";

export interface Account {
  id: number;
  username: string;
  role: AccountRole;
  personnelId?: number | null;
  personnelName?: string | null;
  createdAt: string;
}