# lib/api-zod/src/generated/types/createAccountRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAccountRequest {
  personnelId: number;
  password: string;
}