# lib/api-zod/src/generated/types/updateAccountRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personnel Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAccountRequest {
  password: string;
}